Refactor recursion (#6)
* Use ParamEvaluator class for recursion
